+++ test/pipeline/doc_examples/run_time_dsp/src/main.c
@@ -18,11 +18,11 @@
 void control_thread(adsp_controller_t* control) {
   // convert desired value to parameter type
   float desired_vol_db = -6;
-  int32_t desired_vol_raw = adsp_dB_to_gain(desired_vol_db);
+  int32_t desired_vol_raw = adsp_db_to_gain(desired_vol_db);
 
   adsp_stage_control_cmd_t command = {
     .instance_id = volume_stage_index,
-    .cmd_id = CMD_VOLUME_CONTROL_TARGET_GAIN,
+    .cmd_id = CMD_VOLUME_CONTROL_USER_GAIN,
     .payload_len = sizeof(desired_vol_raw),
     .payload = &desired_vol_raw
   };